% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfa.subtree.R
\name{bfa.subtree}
\alias{bfa.subtree}
\title{Subtree Extractor}
\usage{
bfa.subtree(n, p)
}
\arguments{
\item{n}{subtree size (integer)}

\item{p}{an integer determining the order of bifurcating autoregressive model}
}
\value{
A numeric vector representing a subtree of size \eqn{(2^p-1)} from
the end of a given bifurcating autoregressive tree.
}
\description{
This function extracts subtree of size \eqn{(2^p-1)} from the end of a given
bifurcating autoregressive tree (model) of order p.
}
\examples{
bfa.subtree(31, 1)
bfa.subtree(31, 2)
}
