% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoli.R
\docType{data}
\name{ecoli}
\alias{ecoli}
\title{Lifetimes (in minutes) of lineage E. coli cells.}
\format{
A data frame with 31 rows and 1 variable:
\describe{
\item{lifetime}{E. coli cells lifetime, in minutes}
}
}
\source{
Cowan, R. and Staudte, R. (1986). The Bifurcating Autoregression Model in Cell Lineage
Studies. \emph{Biometrics}, 42(4):769-783.
}
\usage{
ecoli
}
\description{
A real cell lineage dataset taken from Cowan and Staudte (1986).
The values represent the lifetimes in minutes of lineage E. coli cells.
It contains 31 observations making which can be described by a bifurcating
autoregressive tree with 4 generations.
}
\keyword{datasets}
